Move the NEON compatibility defines into a separate file
