+++ arch/arm/slide_hash_neon.c
@@ -9,14 +9,9 @@
  */
 
 #if defined(ARM_NEON_SLIDEHASH)
-#ifdef _M_ARM64
-#  include <arm64_neon.h>
-#else
-#  include <arm_neon.h>
-#endif
+#include "neon_intrins.h"
 #include "../../zbuild.h"
 #include "../../deflate.h"
-#include "../../fallback_builtins.h"
 
 /* SIMD version of hash_chain rebase */
 static inline void slide_hash_chain(Pos *table, uint32_t entries, uint16_t wsize) {
